requirements.txt
djangorestframework==3.9.2
Django==2.2.13
